Updated aludels + ash for Simplified Chinese
I think 容器 is more descriptive for the aludels. 
I'm going back to 灰 for "ash." 粉末 to me sounds more like a finer dusty powder, like makeup or  something, which works for the alchemical dusts, but not for ash, imo. 
Should we get the "alchemical" aspect in the name somewhere? Maybe 蔚蓝炼金粉末? Is that too long?
Thanks to crafteverywhere for improving a lot of my horrid translations and for 梨形 making me laugh. XD