added some recipes
Added the following recipes:

4 Iron Block -> 9 Ender Pearl
2 Ender Pearl -> 1 Gold Ingot
8 Ender Pearl -> 1 Diamond
4 Clay Ball -> 1 Obsidian* (with transmutationStone also in crafting grid)

1 Furnace -> 8 Cobblestone
1 Flower Pot -> 3 Brick
1 Brick -> 1 Clay Ball
1 Chest -> 2 Wood Log
1 Fence Gate -> 1 Wood Log
1 Wooden Door -> 6 Wood Planks
1 Crafting Table -> 4 Wood Planks
1 Fence -> 3 Stick
1 Cauldron -> 7 Iron Ingot
1 Iron Door -> 6 Iron Ingot
1 Minecart -> 5 Iron Ingot
1 Bucket -> 3 Iron Ingot
1 Bread -> 3 Wheat
1 Nether Brick Fence -> 1 Nether Brick Block (netherFence, netherBrick)
Yeah, I said Nether Brick BLOCK there so it doesn't get confused with the Nether Brick counterpart of the regular Brick introduced in the Minecraft 1.5 update.

All of these recipes can be proven equivalent exchanges with a little algebra work on the existing recipes. In fact, I would like to add some more, but I have not done so. In each case, it is due to one or more of the following reasons:
It conflicts with an existing recipe.
It presents some sort of problem.
I literally don't know how to add it.

Most of these miscellaneous other proposals are additions to existing equivalency lists. My most controversial ideas can be found here. Even if implemented, one will almost definitely get nerfed in Pre2. Read on to see what I mean.
Some are deconstruction recipes. One of these introduces another debatable concept.
The last one is a smelting recipe that would probably make a certain minority of house builders very, very happy. You could call it slightly O-P, but for what it is, I think it's okay. But more on that later.
Here's the list:

Sticks should be equivalent to Gravel, Sandstone, and Flint. Manual miners would probably find this quite useful, as they will almost always have too much gravel and not enough torches. In my experience, that's a pretty inevitable and annoying situation.
Sugar Cane, Paper, and Sugar should all be equivalent 1:1. However, the pulverizer from Thermal Expansion allows you to get two Sugar per Sugar Cane. So, while it might be nice sometimes if this recipe was added, it would also add more TE+EE3 dupe loop goop.
Clay Block and Obsidian should definitely be equivalent. Right now the best you can do is 4 Obsidian -> 1 Iron Ingot -> 4 Clay Block, and you can't go the other way. As a workaround to make the other way possible, I added a 4 Clay Ball -> 1 Obsidian recipe. I'd add the code necessary to make them transmutable in-world, but I don't know how to do that.
Wool should definitely be deconstructable, but the color cycler gets in the way.
Cobblestone Wall and Mossy Cobblestone Wall should be equivalent just like Stone Brick and Mossy Stone Brick. However, they should NOT be equivalent to fences, as that would introduce a duplication loophole in EE3 alone.
Speaking of which, we should be able to deconstruct those walls for Cobblestone or Moss Stone 1:1. But wait a second, that introduces a Cobblestone -> Moss Stone conversion. Should we have that? Why or why not? I'll leave you guys to debate that one, because honestly I have no idea if that should be allowed.
Should buttons be equivalent to the blocks used to make them? Again, I am undecided.
Should Carrots be equivalent to Potatoes? Tough call. While every single NPC Village I have seen has both, a carrot is worth 4 times as much hunger as a potato (2 bars vs. 1/2 of a bar). On top of that, carrots are worth 4.8 points of saturation, 8 times as much as the potato's meager 0.6 points. But then one could argue that the Carrot -> Potato conversion is just as valuable, because a BAKED Potato is even better than a Carrot. 50% better on both visible hunger and saturation, to be precise. Hmmm... Carrot and Potato equivalency... a very tough call indeed...
And now for my last idea (but just for the moment and last but not least of course). Seeing as a Clay Ball makes a Brick when you smelt it and 4 Bricks make a Brick Block, shouldn't we be able to just skip all that nonsense and directly smelt a Clay Block into a Brick Block? Well shoot, why the heck not? I'm sure I'm not the only one who has ever wanted to build a brick house, but just doesn't have that sort of patience. This recipe could fix that for us. It would be perfect. Think about it. Do the math. 3 diamonds, 55 coal, and a door would be all you'd need to build that solid brick single-door 9 by 9 of your dreams, floor and all with 1 Brick Block to spare.

I hope I've been of good help here. Given that this is my first pull request outside of a tiny typo fix, I welcome any comments and advice as I learn how to be a good and productive GitHub user. I will also admit that while I now have almost a year of C programming classes under my belt, I know next to nothing specifically about Java programming. I have done my best in my double- and triple-checks to make sure I followed what I assume is a universal recipe template, but feel free to correct me if I am wrong in this assumption and/or point out any mistakes that may have slipped through.

Finally, if Pahimar approves my changes, I hope that people like these new recipes and find them useful. Again, any feedback and especially any advice is greatly appreciated, but constructive criticism only please. Thank you!