Fix for NPE with Mystcraft Writing desk
Fix for https://github.com/pahimar/Equivalent-Exchange-3/issues/348

The Mystcraft writing desk apparently does not have an item registered with the same id as the block. This causes a null pointer exception to occur when you call getItemDamage() on the item stack made from hovering over the writing desk with a Minium stone.