and then, pt_BR.lang was born.
I'd like to know a little more about the keys, because I don't really know what it's about. I really just assumed "release" was about the mod's version, so if it's wrong, let me know.
Other than that, it's all fine. not much to be said, actually, so I'll just stop typing...